+++ __init__.py
@@ -4,7 +4,7 @@ bl_info = {
     "name": "Storytools - Storyboard Tools",
     "description": "Set of tools for Storyboarding",
     "author": "Samuel Bernou",
-    "version": (2, 0, 3),
+    "version": (2, 1, 0),
     "blender": (4, 3, 0),
     "location": "View3D",
     "warning": "",